refactor: group unicode back again
